move the admire button into its own component to prepare for the big remove admire tomorrow
