+++ src/components/Feed/Socialize/FeedEventSocializeSection.tsx
@@ -1,25 +1,22 @@
 import { useFragment } from 'react-relay';
-import { ConnectionHandler, graphql } from 'relay-runtime';
+import { graphql } from 'relay-runtime';
 import { FeedEventSocializeSectionFragment$key } from '__generated__/FeedEventSocializeSectionFragment.graphql';
 import { HStack, VStack } from 'components/core/Spacer/Stack';
-import { AdmireIcon, CommentIcon } from 'icons/SocializeIcons';
+import { CommentIcon } from 'icons/SocializeIcons';
 import styled from 'styled-components';
 import { CommentBox } from 'components/Feed/Socialize/CommentBox';
 import { useCallback, useRef, useState } from 'react';
 import { Interactions } from 'components/Feed/Socialize/Interactions';
 import { FeedEventSocializeSectionQueryFragment$key } from '../../../../__generated__/FeedEventSocializeSectionQueryFragment.graphql';
-import { usePromisifiedMutation } from 'hooks/usePromisifiedMutation';
-import { FeedEventSocializeSectionAdmireMutation } from '../../../../__generated__/FeedEventSocializeSectionAdmireMutation.graphql';
 import {
   FEED_EVENT_ROW_WIDTH_DESKTOP,
   FEED_EVENT_ROW_WIDTH_TABLET,
 } from 'components/Feed/dimensions';
 import breakpoints from 'components/core/breakpoints';
-import { useToastActions } from 'contexts/toast/ToastContext';
-import { AdditionalContext, useReportError } from 'contexts/errorReporting/ErrorReportingContext';
 import { useModalActions } from 'contexts/modal/ModalContext';
 import { AuthModal } from 'hooks/useAuthModal';
 import colors from 'components/core/colors';
+import { AdmireButton } from 'components/Feed/Socialize/AdmireButton';
 
 type FeedEventSocializeSectionProps = {
   onPotentialLayoutShift: () => void;
@@ -35,12 +32,9 @@ export function FeedEventSocializeSection({
   const event = useFragment(
     graphql`
       fragment FeedEventSocializeSectionFragment on FeedEvent {
-        id
-        dbid
-        hasViewerAdmiredEvent
-
         ...CommentBoxFragment
         ...InteractionsFragment
+        ...AdmireButtonFragment
       }
     `,
     eventRef
@@ -60,6 +54,7 @@ export function FeedEventSocializeSection({
           }
         }
 
+        ...AdmireButtonQueryFragment
         ...InteractionsQueryFragment
         ...CommentBoxQueryFragment
         ...useAuthModalFragment
@@ -68,24 +63,6 @@ export function FeedEventSocializeSection({
     queryRef
   );
 
-  const [admire] = usePromisifiedMutation<FeedEventSocializeSectionAdmireMutation>(graphql`
-    mutation FeedEventSocializeSectionAdmireMutation($eventId: DBID!, $connections: [ID!]!)
-    @raw_response_type {
-      admireFeedEvent(feedEventId: $eventId) {
-        ... on AdmireFeedEventPayload {
-          __typename
-          admire @prependNode(edgeTypeName: "FeedEventAdmireEdge", connections: $connections) {
-            dbid
-            ...AdmireLineFragment
-            ...AdmireNoteFragment
-          }
-        }
-      }
-    }
-  `);
-
-  const reportError = useReportError();
-  const { pushToast } = useToastActions();
   const { showModal } = useModalActions();
 
   const commentIconRef = useRef<HTMLDivElement | null>(null);
@@ -107,111 +84,6 @@ export function FeedEventSocializeSection({
     setShowCommentBox((previous) => !previous);
   }, [query, showModal]);
 
-  const handleAdmire = useCallback(async () => {
-    if (query.viewer?.__typename !== 'Viewer') {
-      showModal({
-        content: <AuthModal queryRef={query} />,
-      });
-
-      return;
-    }
-
-    const errorMetadata: AdditionalContext['tags'] = {
-      eventId: event.dbid,
-    };
-
-    function pushErrorToast() {
-      pushToast({
-        autoClose: true,
-        message: `Something went wrong while admiring this post. We're actively looking into it.`,
-      });
-    }
-
-    try {
-      const interactionsConnection = ConnectionHandler.getConnectionID(
-        event.id,
-        'Interactions_admires'
-      );
-      const notesModalConnection = ConnectionHandler.getConnectionID(
-        event.id,
-        'NotesModal_interactions'
-      );
-
-      const optimisticAdmireId = Math.random().toString();
-      const response = await admire({
-        updater: (store, response) => {
-          if (response.admireFeedEvent?.__typename === 'AdmireFeedEventPayload') {
-            const pageInfo = store.get(interactionsConnection)?.getLinkedRecord('pageInfo');
-
-            pageInfo?.setValue(((pageInfo?.getValue('total') as number) ?? 0) + 1, 'total');
-
-            store.get(event.id)?.setValue(true, 'hasViewerAdmiredEvent');
-          }
-        },
-        optimisticResponse: {
-          admireFeedEvent: {
-            __typename: 'AdmireFeedEventPayload',
-            admire: {
-              __typename: 'Admire',
-              id: `Admire:${optimisticAdmireId}`,
-              dbid: optimisticAdmireId,
-              creationTime: new Date().toISOString(),
-              admirer: {
-                id: query.viewer?.user?.id ?? 'unknown',
-                dbid: query.viewer?.user?.dbid ?? 'unknown',
-                username: query.viewer?.user?.username ?? null,
-              },
-            },
-          },
-        },
-        variables: {
-          eventId: event.dbid,
-          connections: [interactionsConnection, notesModalConnection],
-        },
-      });
-
-      if (response.admireFeedEvent?.__typename === 'AdmireFeedEventPayload') {
-        // Tell the virtualized list that some data has changed
-        // therefore this cell's height might change.
-        //
-        // Ideally, this lives in a useEffect inside of the
-        // changing data's component, but right now the virtualized
-        // list is remounting the component every update, causing
-        // an infinite useEffect to occur
-        setTimeout(() => {
-          onPotentialLayoutShift();
-        }, 100);
-      } else {
-        pushErrorToast();
-        reportError(
-          `Could not post comment on feed event, typename was ${response.admireFeedEvent?.__typename}`,
-          {
-            tags: errorMetadata,
-          }
-        );
-      }
-    } catch (error) {
-      pushErrorToast();
-
-      if (error instanceof Error) {
-        reportError(error);
-      } else {
-        reportError(`Could not post comment on feed event for an unknown reason`, {
-          tags: errorMetadata,
-        });
-      }
-    }
-  }, [
-    admire,
-    event.dbid,
-    event.id,
-    onPotentialLayoutShift,
-    pushToast,
-    query,
-    reportError,
-    showModal,
-  ]);
-
   return (
     <SocializedSectionPadding>
       <SocializeSectionWrapper>
@@ -222,7 +94,11 @@ export function FeedEventSocializeSection({
 
           <HStack align="center">
             <IconWrapper>
-              <AdmireIcon onClick={handleAdmire} active={event.hasViewerAdmiredEvent ?? false} />
+              <AdmireButton
+                eventRef={event}
+                queryRef={query}
+                onPotentialLayoutShift={onPotentialLayoutShift}
+              />
             </IconWrapper>
 
             <IconWrapper>
